+++ src/main/java/gwt/jelement/bluetooth/BluetoothUUID.java
@@ -24,24 +24,24 @@ import jsinterop.annotations.JsType;
 public class BluetoothUUID{
     
     @JsMethod(name = "canonicalUUID")
-    public native String canonicalUUID(long alias);
+    public native String canonicalUUID(double alias);
     
     @JsMethod(name = "getCharacteristic")
     public native String getCharacteristic(String name);
     
     @JsMethod(name = "getCharacteristic")
-    public native String getCharacteristic(long name);
+    public native String getCharacteristic(double name);
     
     @JsMethod(name = "getDescriptor")
     public native String getDescriptor(String name);
     
     @JsMethod(name = "getDescriptor")
-    public native String getDescriptor(long name);
+    public native String getDescriptor(double name);
     
     @JsMethod(name = "getService")
     public native String getService(String name);
     
     @JsMethod(name = "getService")
-    public native String getService(long name);
+    public native String getService(double name);
     
 }
