+++ src/main/java/gwt/jelement/css/CSSGroupingRule.java
@@ -36,9 +36,9 @@ public class CSSGroupingRule extends CSSRule{
     public native CSSRuleList getCssRules();
     
     @JsMethod(name = "deleteRule")
-    public native void deleteRule(long index);
+    public native void deleteRule(double index);
     
     @JsMethod(name = "insertRule")
-    public native long insertRule(String rule, long index);
+    public native double insertRule(String rule, double index);
     
 }
