+++ src/main/java/gwt/jelement/dom/DocumentFragment.java
@@ -44,7 +44,7 @@ public class DocumentFragment extends Node{
     public native Element getLastElementChild();
     
     @JsProperty(name="childElementCount")
-    public native long getChildElementCount();
+    public native double getChildElementCount();
     
     @JsMethod(name = "append")
     public native void append(Node nodes);
