+++ src/main/java/gwt/jelement/dom/DocumentOrShadowRoot.java
@@ -40,10 +40,10 @@ public class DocumentOrShadowRoot{
     public native Element getFullscreenElement();
     
     @JsMethod(name = "elementFromPoint")
-    public native Element elementFromPoint(long x, long y);
+    public native Element elementFromPoint(double x, double y);
     
     @JsMethod(name = "elementsFromPoint")
-    public native Element[] elementsFromPoint(long x, long y);
+    public native Element[] elementsFromPoint(double x, double y);
     
     @JsMethod(name = "getSelection")
     public native Selection getSelection();
